57, 65 - optimizations (bs, lr), bug fix, experiments
1) Fixes 65 bug:
Changes learning rate for the default configuration that uses BERTModel
such that it is low and the model doesn't diverge far.
Bug: the model outputs the same logits no matter what input.
Refs:
- https://github.com/huggingface/transformers/issues/735
- https://stackoverflow.com/questions/61855486/bert-encoding-layer-produces-same-output-for-all-inputs-during-evaluation-pytor

2) 57 - Optimizations to run on multi-GPU, to allow auto batch size finder,

https://app.asana.com/0/1204113211080833/1205506702107139/f

3) 45 - Contains code for BrainScore integration but it is commented out

https://app.asana.com/0/1204113211080833/1205278577974214/f

Adapter to calculate brain scores each time after fine tuning. Commented
out so we can merge to branches independently. This also adds a field
for ds000212 to allow brain score calculations.

4) Scripts to run experiments.

https://app.asana.com/0/1204113211080833/1204398674792942/f